MA-18243[android-8mq] Set the GRALLOC private bit according to the bits in display port
To support tiled framebuffer and linear framebuffer at the same time according
to display driver support tile framebuffer or not, use the Bit6 of the display
port byte to indicate that if the display driver support tiled framebuffer.
When Bit6 of display port is 1, it means this display do not support tiled
framebuffer as hardware limitation.
This setting only affect the imx8mq.

Change-Id: Ia72fbeded3ce51bcec36f8f1d803fb491694b7f6
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15104
Reviewed-by: Xuegang Liu <xuegang.liu@nxp.com>